refactor: players 분리
